QuestDBSender

Documentation for QuestDBSender

QuestDBClient.QuestDBSender.DEFAULT_BUFFER_SIZE

QuestDBClient.QuestDBSender.Auth

QuestDBClient.QuestDBSender.Sender

QuestDBClient.QuestDBSender.clear

QuestDBClient.QuestDBSender.close

QuestDBClient.QuestDBSender.connect

QuestDBClient.QuestDBSender.flush

QuestDBClient.QuestDBSender.write

QuestDBClient.QuestDBSender.DEFAULT_BUFFER_SIZE — Constant

const DEFAULT_BUFFER_SIZE::Int = 128 * 1024

Default buffer size of the sender's buffer

QuestDBClient.QuestDBSender.Auth — Type

See: https://questdb.io/docs/reference/api/ilp/authenticate

Authentication object used

QuestDBClient.QuestDBSender.Sender — Type

@with_kw mutable struct Sender
    host::String="127.0.0.1"
    port::Int=9009
    batchSize::Int=DEFAULT_BUFFER_SIZE
    tlsMode::Bool=false
    hasTable::Bool=false
    hasFields::Bool=false
    auth::Union{Auth, Nothing} = nothing ## Can be changed into a new 
    buffer::String = ""
    socket::Union{TCPSocket, Nothing} = nothing
end

Sender struct is the entity responsible for connecting to the QuestDB Server, build records & send them using the ILP protocol.

Arguments

host::String - the host address. The default one is the localhost or 127.0.0.1

port::Int - the port connected to on the host machine. Default value is 9009

batchSize::Int - the buffer size beyond which the contents of the buffer written to the server. Default size is 128 * 1024

hasTable::Bool - used to indicate if an ILP record statement has a table defined

hasFields::Bool - used to indicate if an ILP record statement has fields defined

buffer::String - used to buffer ILP record statements before writing them to the server

socket::Union{TCPSocket, Nothing} - holds the socket connection to the QuestDB Server instance

QuestDBClient.QuestDBSender.write — Method

write(sender::Sender)

Attempts to write the ILP record statements to the server. In case of an error, an exception is thrown

QuestDBClient.QuestDBSender.clear — Method

clear(sender::Sender)

Clears the buffer contents

QuestDBClient.QuestDBSender.connect — Method

connect(sender::Sender)

Attempts to connect the sender to the server socket. In case of an error, an exception is thrown

QuestDBClient.QuestDBSender.flush — Method

flush(sender::Sender)

Attempts to flush any unsent text to the server socket. In case of an error, an exception is thrown

QuestDBClient.QuestDBSender.close — Method

close(sender::Sender)

Attempts to close the sender's connection to the server. In case of an error, an exception is thrown